+++ examples/inverse/plot_dics_source_power.py
@@ -64,10 +64,10 @@ noise_csds = csd_epochs(epochs, mode='multitaper', tmin=-0.11,
 stc = dics_source_power(epochs.info, forward, noise_csds, data_csds)
 
 for i, csd in enumerate(data_csds):
-    message = 'DICS source power at %0.1f Hz' % csd.frequencies[0]
+    message = 'DICS source power at %0.1f Hz' % csd.freqs[0]
     brain = stc.plot(surface='inflated', hemi='rh', subjects_dir=subjects_dir,
                      time_label=message, figure=i)
     brain.set_data_time_index(i)
     brain.show_view('lateral')
     # Uncomment line below to save images
-    # brain.save_image('DICS_source_power_freq_%d.png' % csd.frequencies[0])
+    # brain.save_image('DICS_source_power_freq_%d.png' % csd.freqs[0])